добавил ручки для докторов и клиник

+++ src/main/java/ru/mirea/dentalclinic/domain/models/Record.java
@@ -3,11 +3,11 @@ package ru.mirea.dentalclinic.domain.models;
 import java.util.Date;
 
 public record Record(
-        Integer id,
+        Long id,
         Doctor doctor,
         Office office,
-        Long start,
-        Long end,
+        Integer start,
+        Integer end,
         Boolean isBooked,
         Date day
 ) {
